+++ src/icons/logo.tsx
@@ -1,3 +1,25 @@
+export function Logo_md() {
+  return (
+    <svg width="160" height="35" viewBox="0 0 160 35" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M17.856 8.44002V16.9734H11.4773V18.68H17.856V26.36H19.4347V8.44002H17.856ZM8.32001 26.36H9.89867V8.44002H8.32001V26.36Z"
+        fill="#0F172A"/>
+      <path
+        d="M51.584 10.1894C51.2427 9.59202 50.6667 9.14402 49.856 8.84535C49.1093 8.56802 47.9787 8.44002 46.4853 8.44002C44.992 8.44002 43.9467 8.56802 43.1573 8.84535C42.3467 9.14402 41.7707 9.59202 41.408 10.232C41.0453 10.872 40.832 11.8107 40.768 13.0907V26.36H42.3893V18.4027H44.032V16.6747H42.3893V12.8987C42.4533 11.96 42.5813 11.32 42.7947 10.9787C43.1787 10.3387 44.3947 10.0187 46.4853 10.0187C48.576 10.0187 49.8347 10.3387 50.176 10.9574C50.3893 11.2774 50.5173 11.896 50.5813 12.8134V16.6747H45.6533V18.4027H50.5813V26.36H52.1813V13.0267C52.1173 11.7467 51.9253 10.808 51.584 10.1894Z"
+        fill="#0F172A"/>
+      <path
+        d="M85.0133 10.2747C84.672 9.57068 84.032 9.08002 83.0933 8.84535C82.1547 8.58935 80.6613 8.46135 78.6133 8.44002H73.5147V26.36H75.136V18.9147H77.7387V17.2507H75.136V10.0827H80.896C81.8133 10.0827 82.5173 10.232 83.0293 10.5094C83.2853 10.616 83.4987 10.8934 83.648 11.3627C83.7973 11.8107 83.8827 12.6 83.8827 13.688C83.8827 15.5014 83.6053 16.5467 83.0293 16.8454C82.6027 17.1014 81.8987 17.2294 80.96 17.2507H79.0613V18.8934H81.024L82.176 18.7867L82.6453 18.68L83.2213 18.4667C83.9893 18.232 84.5653 17.7627 84.9493 17.0587C85.3333 16.376 85.5253 15.2454 85.5253 13.688C85.5253 12.1307 85.3547 11 85.0133 10.2747Z"
+        fill="#0F172A"/>
+      <path
+        d="M113.749 8.45068C111.723 8.47202 110.251 8.60002 109.312 8.83468C108.373 9.06935 107.755 9.56002 107.392 10.264C107.051 10.9894 106.859 12.12 106.859 13.6774V21.1227C106.859 22.7014 107.051 23.832 107.435 24.5147C107.84 25.1974 108.395 25.688 109.141 25.944L109.76 26.136L110.229 26.2427L111.403 26.3494H117.781V24.728H111.424C110.421 24.6427 109.717 24.4934 109.355 24.3014C108.779 23.9814 108.48 22.9147 108.48 21.1227V18.2854H114.048V16.536H108.48V13.6774C108.48 11.864 108.757 10.7974 109.312 10.4987C109.696 10.3067 110.315 10.1787 111.147 10.1147H111.488L117.781 10.0934V8.45068H113.749ZM115.925 18.2854H117.781V16.536H115.925V18.2854Z"
+        fill="#0F172A"/>
+      <path
+        d="M148.117 18.7227L148.288 18.68L148.843 18.4667C149.611 18.2107 150.187 17.7414 150.549 17.0587C150.933 16.376 151.125 15.2454 151.125 13.6667C151.125 12.088 150.912 10.872 150.507 10.1254C150.101 9.40002 149.461 8.93068 148.587 8.73868C147.712 8.54668 146.347 8.44002 144.491 8.44002H139.115V26.36H140.757V18.9147H145.259V17.272H140.757V10.0827H146.88C147.755 10.1894 148.373 10.3174 148.693 10.5094C149.248 10.7867 149.525 11.832 149.525 13.6667C149.525 14.8187 149.44 15.608 149.269 16.0134C149.077 16.4187 148.885 16.696 148.672 16.8454C148.245 17.08 147.563 17.2294 146.624 17.272H146.432V18.8934L149.909 26.36H151.68L148.117 18.7227Z"
+        fill="#0F172A"/>
+    </svg>
+  )
+}
+
 export function Logo_lg() {
   return (
     <svg width="200" height="44" viewBox="0 0 200 44" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -43,3 +65,4 @@ export function Logo_lg() {
     </svg>
   )
 }
+
